defog, healing wish, trick room, magic room, wonder room

+++ src/include/battle_locations.h
@@ -9,7 +9,7 @@ u16 battle_executed_move;
 u8 battle_attacker_bank;
 u8 battle_defender_bank;
 u8 battle_damage_multiplier;
-u16 battle_damage_store;
+s32 battle_damage_store;
 u32 battle_hitmarker;
 void* battlescript_cursor;
 u8 battle_menu_chosen_item[4];
